requirements.txt
Flask==2.1.0
gunicorn==20.1.0

# for CF
pandas==1.3.5
numpy==1.21.6
matplotlib==3.2.2
seaborn==0.11.2
sklearn-pandas==1.8.0
requests==2.23.0